TIP-843: Move product array converter spec to enrich
