+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/ArrayConverter/FlatToStandard/AssociationColumnsResolverSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
