+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/ArrayConverter/FlatToStandard/FieldSplitterSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard;
 
 use PhpSpec\ObjectBehavior;
