+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/ArrayConverter/FlatToStandard/ValueConverter/SimpleSelectConverterSpec.php
@@ -1,23 +1,16 @@
 <?php
 
-namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\ValueConverter;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\FieldSplitter;
 
-class TextConverterSpec extends ObjectBehavior
+class SimpleSelectConverterSpec extends ObjectBehavior
 {
     function let(FieldSplitter $fieldSplitter)
     {
-        $this->beConstructedWith(
-            $fieldSplitter,
-            [
-                'pim_catalog_identifier',
-                'pim_catalog_text',
-                'pim_catalog_textarea'
-            ]
-        );
+        $this->beConstructedWith($fieldSplitter, ['pim_catalog_simpleselect']);
     }
 
     function it_is_a_converter()
@@ -27,39 +20,21 @@ class TextConverterSpec extends ObjectBehavior
 
     function it_supports_converter_field()
     {
-        $this->supportsField('pim_catalog_identifier')->shouldReturn(true);
-        $this->supportsField('pim_catalog_text')->shouldReturn(true);
-        $this->supportsField('pim_catalog_textarea')->shouldReturn(true);
-        $this->supportsField('pim_catalog_price')->shouldReturn(false);
+        $this->supportsField('pim_catalog_simpleselect')->shouldReturn(true);
+        $this->supportsField('pim_catalog_identifier')->shouldReturn(false);
     }
 
-    function it_converts_a_string(AttributeInterface $attribute)
+    function it_converts(AttributeInterface $attribute)
     {
         $attribute->getCode()->willReturn('attribute_code');
         $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
 
-        $value = 'my awesome text';
+        $value = 'my_awesome_identifier';
 
         $expectedResult = ['attribute_code' => [[
             'locale' => 'en_US',
             'scope'  => 'mobile',
-            'data'   => 'my awesome text',
-        ]]];
-
-        $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
-    }
-
-    function it_casts_a_number_to_string_during_conversion(AttributeInterface $attribute)
-    {
-        $attribute->getCode()->willReturn('attribute_code');
-        $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
-
-        $value = 1234;
-
-        $expectedResult = ['attribute_code' => [[
-            'locale' => 'en_US',
-            'scope'  => 'mobile',
-            'data'   => '1234',
+            'data'   => 'my_awesome_identifier',
         ]]];
 
         $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
