+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/ArrayConverter/FlatToStandard/ValueConverter/TextConverterSpec.php
@@ -1,16 +1,23 @@
 <?php
 
-namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\ValueConverter;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\FieldSplitter;
 
-class NumberConverterSpec extends ObjectBehavior
+class TextConverterSpec extends ObjectBehavior
 {
     function let(FieldSplitter $fieldSplitter)
     {
-        $this->beConstructedWith($fieldSplitter, ['pim_catalog_number']);
+        $this->beConstructedWith(
+            $fieldSplitter,
+            [
+                'pim_catalog_identifier',
+                'pim_catalog_text',
+                'pim_catalog_textarea'
+            ]
+        );
     }
 
     function it_is_a_converter()
@@ -20,14 +27,31 @@ class NumberConverterSpec extends ObjectBehavior
 
     function it_supports_converter_field()
     {
-        $this->supportsField('pim_catalog_number')->shouldReturn(true);
+        $this->supportsField('pim_catalog_identifier')->shouldReturn(true);
+        $this->supportsField('pim_catalog_text')->shouldReturn(true);
+        $this->supportsField('pim_catalog_textarea')->shouldReturn(true);
         $this->supportsField('pim_catalog_price')->shouldReturn(false);
     }
 
-    function it_converts_an_integer(AttributeInterface $attribute)
+    function it_converts_a_string(AttributeInterface $attribute)
     {
         $attribute->getCode()->willReturn('attribute_code');
-        $attribute->isDecimalsAllowed()->willReturn(false);
+        $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
+
+        $value = 'my awesome text';
+
+        $expectedResult = ['attribute_code' => [[
+            'locale' => 'en_US',
+            'scope'  => 'mobile',
+            'data'   => 'my awesome text',
+        ]]];
+
+        $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
+    }
+
+    function it_casts_a_number_to_string_during_conversion(AttributeInterface $attribute)
+    {
+        $attribute->getCode()->willReturn('attribute_code');
         $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
 
         $value = 1234;
@@ -35,41 +59,23 @@ class NumberConverterSpec extends ObjectBehavior
         $expectedResult = ['attribute_code' => [[
             'locale' => 'en_US',
             'scope'  => 'mobile',
-            'data'   => 1234,
+            'data'   => '1234',
         ]]];
 
         $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
     }
 
-    function it_converts_an_integer_formatted_as_string(AttributeInterface $attribute)
+    function it_returns_null_data_if_empty_value_provided(AttributeInterface $attribute)
     {
         $attribute->getCode()->willReturn('attribute_code');
-        $attribute->isDecimalsAllowed()->willReturn(false);
         $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
 
-        $value = '1234';
+        $value = '';
 
         $expectedResult = ['attribute_code' => [[
             'locale' => 'en_US',
             'scope'  => 'mobile',
-            'data'   => 1234,
-        ]]];
-
-        $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
-    }
-
-    function it_does_not_convert_a_float(AttributeInterface $attribute)
-    {
-        $attribute->getCode()->willReturn('attribute_code');
-        $attribute->isDecimalsAllowed()->willReturn(true);
-        $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
-
-        $value = 1234.36;
-
-        $expectedResult = ['attribute_code' => [[
-            'locale' => 'en_US',
-            'scope'  => 'mobile',
-            'data'   => 1234.36,
+            'data'   => null,
         ]]];
 
         $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
