--- src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/Product/ValueConverter/SimpleSelectConverterSpec.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter;
-
-use PhpSpec\ObjectBehavior;
-use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\FieldSplitter;
-
-class SimpleSelectConverterSpec extends ObjectBehavior
-{
-    function let(FieldSplitter $fieldSplitter)
-    {
-        $this->beConstructedWith($fieldSplitter, ['pim_catalog_simpleselect']);
-    }
-
-    function it_is_a_converter()
-    {
-        $this->shouldImplement('Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter\ValueConverterInterface');
-    }
-
-    function it_supports_converter_field()
-    {
-        $this->supportsField('pim_catalog_simpleselect')->shouldReturn(true);
-        $this->supportsField('pim_catalog_identifier')->shouldReturn(false);
-    }
-
-    function it_converts(AttributeInterface $attribute)
-    {
-        $attribute->getCode()->willReturn('attribute_code');
-        $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
-
-        $value = 'my_awesome_identifier';
-
-        $expectedResult = ['attribute_code' => [[
-            'locale' => 'en_US',
-            'scope'  => 'mobile',
-            'data'   => 'my_awesome_identifier',
-        ]]];
-
-        $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
-    }
-
-    function it_returns_null_data_if_empty_value_provided(AttributeInterface $attribute)
-    {
-        $attribute->getCode()->willReturn('attribute_code');
-        $fieldNameInfo = ['attribute' => $attribute, 'locale_code' => 'en_US', 'scope_code' => 'mobile'];
-
-        $value = '';
-
-        $expectedResult = ['attribute_code' => [[
-            'locale' => 'en_US',
-            'scope'  => 'mobile',
-            'data'   => null,
-        ]]];
-
-        $this->convert($fieldNameInfo, $value)->shouldReturn($expectedResult);
-    }
-}
